requirement.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
opencv-python==4.11.0.86
pygame-ce==2.5.3
pygame_gui==0.6.13
python-i18n==0.3.9
shapely==2.0.7
Werkzeug==3.1.3
